finding largest number of an array
it coded in my logic, that is i take first element(zero index) as largest and then it is compared to the rest of the array elements. if any element is found to be greater than the previously taken the value is swapped. and so, by the end of the loop we get largest element.